approaches/explodingkittens/EK Code v2/requirements.txt
albumentations
joblib
lightning
matplotlib
nibabel
numpy
opencv-python
pandas
scikit-learn
scipy
timm
torchvision
tqdm